Moves models to separate folder
