+++ src/components/Block.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import './Block.css';
-import Cell from '../utils/Cell';
+import Cell from '../models/Cell';
 
 export interface BlockProps {
     cell: Cell;
@@ -9,8 +9,8 @@ export interface BlockProps {
 function Block(props: BlockProps) {
     const { value, position } = props.cell;
     const style = {
-        "--position-x" : position.x,
-        "--position-y" : position.y
+        "--position-x": position.x,
+        "--position-y": position.y
     } as React.CSSProperties;
     return <div style={style} className={`block block_value_${value}`}>
         <span className={`block__text block__text_digits-count_${value.toString().length}`}>
