+++ src/models/Actions/AddCellAction.ts
@@ -2,9 +2,9 @@ import Action from './Action';
 import Cell from '../Cell';
 
 class AddCellAction extends Action {
-    cell : Cell;
+    cell: Cell;
 
-    constructor(cell : Cell) {
+    constructor(cell: Cell) {
         super();
         this.cell = cell;
     }
