+++ src/models/Actions/LoseAction.ts
