+++ src/models/Cell.ts
@@ -2,10 +2,10 @@ import Point from './Point';
 import ICloneable from './ICloneable';
 
 class Cell implements ICloneable {
-    value : number;
-    position : Point;
+    value: number;
+    position: Point;
 
-    constructor(value : number, position : Point) {
+    constructor(value: number, position: Point) {
         this.value = value;
         this.position = position;
     }
