+++ src/models/ICloneable.ts
