+++ src/models/Logic/ILogic.ts
