+++ src/models/LogicRandomize/ILogicRandomize.ts
