+++ src/models/Point.ts
